requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
 # Built-in with Python

# AI/ML dependencies (for production RAG system)
# Uncomment when implementing production features:
openai==1.3.5  # For OpenAI GPT models
anthropic==0.7.0  # For Claude API
sentence-transformers==2.2.2  # For embeddings
faiss-cpu==1.7.4  # For vector similarity search
numpy==1.24.3
transformers==4.35.0  # For BERT models

# Utility
python-dotenv==1.0.0
httpx==0.25.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
